Locate analytics events in same file
