+++ src/sentry/analytics/events/__init__.py
@@ -21,11 +21,10 @@ from .first_transaction_sent import *  # noqa: F401,F403
 from .first_user_context_sent import *  # noqa: F401,F403
 from .groupowner_assignment import *  # noqa: F401,F403
 from .inapp_request import *  # noqa: F401,F403
+from .integration_commit_context_all_frames import *  # noqa: F401,F403
 from .integration_failed_to_fetch_commit_context import *  # noqa: F401,F403
-from .integration_failed_to_fetch_commit_context_all_frames import *  # noqa: F401,F403
 from .integration_pipeline_step import *  # noqa: F401,F403
 from .integration_serverless_setup import *  # noqa: F401,F403
-from .integrations_successfully_fetched_commit_context import *  # noqa: F401,F403
 from .internal_integration_created import *  # noqa: F401,F403
 from .issue_alert_fired import *  # noqa: F401,F403
 from .issue_archived import *  # noqa: F401,F403
