+++ src/sentry/analytics/events/integration_commit_context_all_frames.py
@@ -1,7 +1,21 @@
 from sentry import analytics
 
 
-class IntegrationsSuccessfullyFetchedCommitContext(analytics.Event):
+class IntegrationsFailedToFetchCommitContextAllFrames(analytics.Event):
+    type = "integrations.failed_to_fetch_commit_context_all_frames"
+
+    attributes = (
+        analytics.Attribute("organization_id"),
+        analytics.Attribute("project_id"),
+        analytics.Attribute("group_id"),
+        analytics.Attribute("event_id"),
+        analytics.Attribute("num_frames", type=int),
+        analytics.Attribute("num_successfully_mapped_frames", type=int),
+        analytics.Attribute("reason"),
+    )
+
+
+class IntegrationsSuccessfullyFetchedCommitContextAllFrames(analytics.Event):
     type = "integrations.successfully_fetched_commit_context_all_frames"
 
     attributes = (
@@ -19,4 +33,5 @@ class IntegrationsSuccessfullyFetchedCommitContext(analytics.Event):
     )
 
 
-analytics.register(IntegrationsSuccessfullyFetchedCommitContext)
+analytics.register(IntegrationsSuccessfullyFetchedCommitContextAllFrames)
+analytics.register(IntegrationsFailedToFetchCommitContextAllFrames)
